fix(android): resolve level loading RTTI crashes, sky rendering, and menu video transparency
- smart_cast / RTTI: Add RTLD_GLOBAL to dlopen in ModuleLookup and specialize
  smart_cast for IKinematics, IKinematicsAnimated, and IParticleCustom to bypass
  cross-DSO RTTI symbol resolution issues on Android.
- xrRenderGL: Replace fragile PBO video texture streaming with direct CPU uploads
  in glSH_Texture, fix Theora alpha channel decompression clamping, and bind sky
  cubemap textures unconditionally on OpenGL.
- Shaders: Fix GLSL ESSL 3.20 syntax and type mismatches in water shaders (water.ps,
  v_waterd.h) and shadow mapping headers.
- xrUICore / xrGame: Export CUILightAnimColorConroller with XRUICORE_API and add
  defensive null checks in CUIGameLog::Update and physics shell initialization.

+++ res/gamedata/shaders/gl/iostructs/v_waterd.h
@@ -8,7 +8,7 @@ struct v_vert
 	float4 	T	;	// TANGENT;
 	float4 	B	;	// BINORMAL;
 	float4	color	;	// COLOR0;		// (r,g,b,dir-occlusion)
-	int2 	uv	;	// TEXCOORD0;		// (u0,v0)
+	float2 	uv	;	// TEXCOORD0;		// (u0,v0)
 };
 struct v2p
 {
@@ -26,7 +26,7 @@ layout(location = NORMAL)		in float4	v_vert_N		; // NORMAL;		// (nx,ny,nz,hemi o
 layout(location = TANGENT)		in float4	v_vert_T		; // TANGENT;		// tangent
 layout(location = BINORMAL)		in float4	v_vert_B		; // BINORMAL;		// binormal
 layout(location = COLOR0)		in float4	v_vert_color		; // COLOR0;		// (r,g,b,dir-occlusion)
-layout(location = TEXCOORD0)		in int2		v_vert_uv		; // TEXCOORD0;		// (u0,v0)
+layout(location = TEXCOORD0)		in float2	v_vert_uv		; // TEXCOORD0;		// (u0,v0)
 
 layout(location = TEXCOORD0) 		out float2	v2p_vert_tbase		; // TEXCOORD0;
 layout(location = TEXCOORD1) 		out float2	v2p_vert_tdist0		; // TEXCOORD1;
